Revert "moved integration dependent code to new file"
This reverts commit 995f1ea6a2e398bf43f20e8f009593a4283ff899.
